split datatype file into one file per class
