+++ clic/src/core/cleBuffer.cpp
@@ -1,4 +1,4 @@
-#include "cleDataType.hpp"
+#include "cleBuffer.hpp"
 
 namespace cle
 {
